wil6210: change headroom_size type to uint
headroom_size is defined as ushort but accessed using
a module parameter of type uint. When module parameter is
set it will overwrite a nearby variable.
For example on some platforms when running:
modprobe wil6210 rx_large_buf=1 headroom_size=256
After modprobe rx_large_buf will be 0.

+++ wil6210/txrx.c
@@ -53,7 +53,7 @@ MODULE_PARM_DESC(rx_large_buf, " allocate 8KB RX buffers, default - no");
 
 #define WIL6210_MAX_HEADROOM_SIZE	(256)
 
-ushort headroom_size; /* = 0; */
+uint headroom_size; /* = 0; */
 static int headroom_size_set(const char *val, const struct kernel_param *kp)
 {
 	int ret;
